ajustes para que el app arranque
