+++ .devcontainer/devcontainer.json
@@ -6,7 +6,11 @@
 
 	// Features to add to the dev container. More info: https://containers.dev/features.
 	// "features": {}
-
+	"features": {
+        "ghcr.io/devcontainers-extra/features/mvnd-sdkman:2": {
+            "version": "latest"
+        }
+    },
 	// Use 'forwardPorts' to make a list of ports inside the container available locally.
 	// This can be used to network with other containers or with the host.
 	// "forwardPorts": [5432],
