+++ pom.xml
@@ -58,11 +58,18 @@
 		<dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi</artifactId>
+			<version>5.2.5</version>
         </dependency>
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi-ooxml</artifactId>
+			<version>5.2.5</version>
         </dependency>
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+			<scope>annotationProcessor</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
