+++ src/main/resources/application.properties
@@ -1,9 +1,9 @@
 spring.application.name=ats-ventas
-spring.datasource.url=jdbc:postgresql://localhost:5432/ats_ventas
-spring.datasource.username=your_username
-spring.datasource.password=your_password
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:postgresql://dpg-d0vp1sumcj7s73frbr80-a.oregon-postgres.render.com:5432/dbtitulos
+spring.datasource.username=usertitulos
+spring.datasource.password=qEvGwOJvd7l1aL4lWyIDoi60c3tbr0Nx
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
@@ -14,3 +14,5 @@ file.upload-dir=uploads
 server.port=8080
 logging.level.org.springframework=INFO
 logging.level.com.ats_ventas=DEBUG
+
+
